renamed components files as mentor told
